third_party/blink/web_tests/fast/frames/frame-programmatic-noresize-expected.txt



--------
Frame: 'testFrame'
--------


--------
Frame: 'results'
--------
This tests that frame resizing is allowed and disallowed when programmatically removing and adding the noresize attribute, respectively.

Test disallow frame resize via noResize IDL attribute:
PASS document.getElementById("testFrame").getBoundingClientRect().width is 200

Test allow frame resize when noResize IDL attribute has value undefined:
PASS document.getElementById("testFrame").getBoundingClientRect().width is 210

Test allow frame resize when noResize IDL attribute has value null:
PASS document.getElementById("testFrame").getBoundingClientRect().width is 220

Test disallow frame resize via noresize content attribute:
PASS document.getElementById("testFrame").getBoundingClientRect().width is 220
PASS document.getElementById("testFrame").getBoundingClientRect().width is 220
PASS document.getElementById("testFrame").getBoundingClientRect().width is 220
PASS document.getElementById("testFrame").getBoundingClientRect().width is 220

Test disallow then allow frame resize via noResize IDL attribute (i.e. increase frame width by 10 pixels):
PASS document.getElementById("testFrame").getBoundingClientRect().width is 220
PASS document.getElementById("testFrame").getBoundingClientRect().width is 230

Test disallow then allow frame resize via noresize content attribute (i.e. increase frame width by 10 pixels):
PASS document.getElementById("testFrame").getBoundingClientRect().width is 230
PASS document.getElementById("testFrame").getBoundingClientRect().width is 240
